# pubky-homeserver/openapi-client.yml
...
openapi: 3.1.0
info:
  title: Pubky Homeserver Client API
  description: |-
    The Client Server is the primary Pubky Homeserver API for user authentication,
    tenant file storage, and event streaming.

    ## Tenant Identification

    Storage requests at `/storage/{user_z32}/{path}` use the owner in the path,
    while `Host` remains the HTTP authority.

    Deprecated owner-relative routes resolve the owner from `pubky-host`, then
    `Host`, then `?pubky-host=`.

    ## Authentication

    Two authentication methods exist:
    - **Grant-based (current):** `Authorization: Bearer <token>` obtained via `POST /auth/grant/session`.
    - **Cookie-based (deprecated):** Session cookie set by `POST /signup` or `POST /session`.

    A resolved bearer session takes precedence over cookie authentication.

    ## Capabilities

    Private reads and all writes are capability-gated. Capabilities use
    `<scope>:<actions>`, for example `/priv/app/:rw`; `/:rw` covers both roots.

    ## Write Path Restriction

    All PUT/DELETE operations require paths under `/pub/` or `/priv/`. Attempts to
    write elsewhere return `403 Forbidden`.
  version: 0.9.0
  license:
    name: MIT
    identifier: MIT
servers:
- url: http://localhost:6287
  description: Client server (Pubky socket, default port)
security: []
...
paths:
...
  "/session":
    post:
      tags:
      - Auth - Cookie
      summary: Sign in (deprecated)
      operationId: cookieSignin
      deprecated: true
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              description: Postcard-serialized AuthToken
      responses:
        '200':
          description: Signin successful
          headers:
            Set-Cookie:
              description: Session cookie
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid AuthToken
        '404':
          description: User not found
    get:
      tags:
      - Auth - Cookie
      summary: Get current session info (deprecated)
      operationId: getCookieSession
      deprecated: true
      security:
      - cookieAuth: []
      responses:
        '200':
          description: Session info
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: Postcard-serialized legacy session info
        '401':
          description: No valid cookie session
    delete:
      tags:
      - Auth - Cookie
      summary: Logout (deprecated)
      description: Idempotent — returns 200 even without a valid session.
      operationId: cookieSignout
      deprecated: true
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          description: Logged out (idempotent)
          headers:
            Set-Cookie:
              description: Cookie removal header
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Grant-based bearer token from `POST /auth/grant/session`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        Deprecated. In practice, the cookie name is the user's z-base-32
        public key and the value is the session secret. The `name` field here
        is a placeholder; the actual name varies per user.
...